+++ react-course/src/Pages/Home/index.jsx
@@ -1,192 +1,105 @@
-import Card from '../../Components/Card'
-import ProductDetail from '../../Components/ProductDetail'
-import ProductModal from '../../Components/ProductModal'
-import { useShopiContext } from '../../Context'
-import CheckoutSideMenu from '../../Components/CheckoutSideMenu'
+import React from 'react';
+function HomePage (){
 
-import Carousel from 'react-bootstrap/Carousel';
-import Form from 'react-bootstrap/Form';
-import { Container, Row } from 'react-bootstrap'
-import Col from 'react-bootstrap/Col';
-
-function Home() {
-  const { openModal, search, filteredItems,setSearchByCategory,isActiveChocolate,isActiveGomitas,
-    isActiveBotanas, isActiveTodo } = useShopiContext();
-const filtro =(val) =>{
-    setSearchByCategory(val);
-}
-  const renderView = () => {
-      if (filteredItems?.length > 0) {
-        return (
-
-          // <Row xs={2} md={4} className="g-4">
-          //   {Array.from(filteredItems)?.map((item) => (
-          //   <Col  > 
-          //  <Card key={item.id} data={item}></Card>
-          //   </Col>
-          //  ))}
-
-          // </Row>
-
-          filteredItems?.map(item => (
-            <Card key={item.id} data={item}> </Card>
-          ))
+   
+    return(
+        <div  className='w-full '>
 
 
-        )
-      } else {
-        return (
-          <div>No hay resultados</div>
-        )
-      } 
-  }
-  return (
-    <div className="container-fluid ">
-      <Carousel>
-      <Carousel.Item>
-      <img className="w-full h-36 object-cover rounded-lg" src='https://images.pexels.com/photos/209345/pexels-photo-209345.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1' alt='img'></img>
+
+
+<div class="sm:px-8 mt-9"><div class="mx-auto w-full max-w-7xl lg:px-8">
+  <div class="relative px-4 sm:px-8 lg:px-12">
+    <div class="mx-auto max-w-2xl lg:max-w-5xl">
+      <div class="max-w-2xl">
+        <h1 class="text-4xl font-bold tracking-tight text-zinc-800 dark:text-zinc-100 sm:text-5xl">Snacks</h1>
+        <p class="mt-6 text-base text-zinc-600 dark:text-zinc-400">I’m Spencer,ms.</p>
       
-        <Carousel.Caption>
-          <h3>First slide label</h3>
-          <p>Nulla vitae elit libero, a pharetra augue mollis interdum.</p>
-        </Carousel.Caption>
-      </Carousel.Item>
-      <Carousel.Item>
-      <img className="w-full h-36 object-cover rounded-lg" src='https://images.pexels.com/photos/209345/pexels-photo-209345.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1' alt='img'></img>
-        <Carousel.Caption>
-          <h3>Second slide label</h3>
-          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>
-        </Carousel.Caption>
-      </Carousel.Item>
-      <Carousel.Item>
-      <img className="w-full h-36 object-cover rounded-lg" src='https://images.pexels.com/photos/209345/pexels-photo-209345.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1' alt='img'></img>
-        <Carousel.Caption>
-          <h3>Third slide label</h3>
-          <p>
-            Praesent commodo cursus magna, vel scelerisque nisl consectetur.
-          </p>
-        </Carousel.Caption>
-      </Carousel.Item>
-    </Carousel>
-    {/* <Container className=' mt-5 items-center justify-center'>
-    <Row >
-        <Col  md={{ span: 6, offset: 6 }} sm={{ span: 12, offset: 12 }} >
-        <Form.Control
-              type="text"
-              placeholder="Buscar producto"
-              onChange={search}
-              
-            />
-            </Col>
-        </Row>
-    </Container> */}
-       
+                          </div>
+                          </div>
+                          </div>
+                          </div>
+                          </div>
+
+
+<div class="mt-16 sm:mt-20 ">
+  <div class="-my-4 flex justify-center gap-5 overflow-hidden py-4 sm:gap-8">
+    <div class="relative aspect-[9/10] w-44 flex-none overflow-hidden rounded-xl bg-zinc-100 dark:bg-zinc-800 sm:w-72 sm:rounded-2xl rotate-2">
+      <img alt="" loading="lazy" width="3744" height="5616" decoding="async" data-nimg="1" class="absolute inset-0 h-full w-full object-cover" 
+       sizes="(min-width: 640px) 18rem, 11rem" 
+        src="https://media.istockphoto.com/id/1372396798/es/foto/colorido-golosinas.jpg?s=1024x1024&w=is&k=20&c=65qjK3a25y9aP7UT3-5Ab8puj4ZilvtdsdkVM_DLCtI="/>
+       </div>
+       <div class="relative aspect-[9/10] w-44 flex-none overflow-hidden rounded-xl bg-zinc-100 dark:bg-zinc-800 sm:w-72 sm:rounded-2xl -rotate-2">
+        <img alt="" loading="lazy" width="3936" height="2624" decoding="async" data-nimg="1" class="absolute inset-0 h-full w-full object-cover"   src="https://media.istockphoto.com/id/1362087297/es/foto/bocadillos-con-salchichas-queso-manteca-de-cerdo-y-varias-salsas-en-una-boda-ucraniana.jpg?s=2048x2048&w=is&k=20&c=KlfPtZmzH67uAbWfktl7-hcy_rB4i8N1rgYToJgDQc4="/></div><div class="relative aspect-[9/10] w-44 flex-none overflow-hidden rounded-xl bg-zinc-100 dark:bg-zinc-800 sm:w-72 sm:rounded-2xl rotate-2">
+          <img alt="" loading="lazy" width="5760" height="3840" decoding="async" data-nimg="1" class="absolute inset-0 h-full w-full object-cover"  src="https://media.istockphoto.com/id/672552464/es/foto/barra-de-caramelo-de-boda-con-nueces-frutas-y-pasteler%C3%ADa.jpg?s=2048x2048&w=is&k=20&c=kOUGIdZzT7oHN__7RAGacraPeFAByZnH-4hnQBPSPjg="/></div><div class="relative aspect-[9/10] w-44 flex-none overflow-hidden rounded-xl bg-zinc-100 dark:bg-zinc-800 sm:w-72 sm:rounded-2xl rotate-2">
+            <img alt="" loading="lazy" width="2400" height="3000" decoding="async" data-nimg="1" class="absolute inset-0 h-full w-full object-cover"   src="https://media.istockphoto.com/id/513943132/es/foto/dulces-surtidos-sobre-la-mesa.jpg?s=2048x2048&w=is&k=20&c=ffaI83Y9fomrsdGm48b04WtfUhD4T9agJi7otKQBS_4="/></div><div class="relative aspect-[9/10] w-44 flex-none overflow-hidden rounded-xl bg-zinc-100 dark:bg-zinc-800 sm:w-72 sm:rounded-2xl -rotate-2">
+              <img alt="" loading="lazy" width="4240" height="2384" decoding="async" data-nimg="1" class="absolute inset-0 h-full w-full object-cover"   src="https://media.istockphoto.com/id/178565437/es/foto/yogur-helado-ingredientes.jpg?s=2048x2048&w=is&k=20&c=3FeQk1y0TvNo4EY-VjJgifsOW1UCUJ2-qw2orrPQFsE="/></div></div></div>
 
 
 
 
-    <div className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8">
 
 
-    <div className="mt-4 grid grid-cols-1 items-start gap-x-8 gap-y-8 sm:mt-16 sm:gap-y-16 lg:mt-24 lg:grid-cols-4">
-      <div className="relative -mx-4 flex overflow-x-auto pb-4 sm:mx-0 sm:block sm:overflow-visible sm:pb-0">
-    {/* <h2 id="speakers-title" className="font-display text-4xl font-medium tracking-tighter text-blue-600 sm:text-5xl">Speakers</h2> 
-      <p className="mt-4 font-display text-2xl tracking-tight text-blue-900">
-        Learn from the experts on the cutting-edge of deception at the most sinister companies.
-        </p> */}
-      
-      </div>
-      <div className="lg:col-span-3">
-      <div className='mt-4 font-display text-2xl tracking-tight text-blue-900'>
-        <input type='text' placeholder='Buscar producto' className='rounded-lg border border-green-200 shadow p-3 mb-3 focus:outline-none' 
-          onChange={search}></input>
-      </div>
-        </div>
-</div>
+<div className="sticky top-0 z-50"><div className="sm:hidden" data-headlessui-state="">
+  <div className="relative flex items-center px-4 py-3 bg-white/95 shadow-sm [@supports(backdrop-filter:blur(0))]:bg-white/80 [@supports(backdrop-filter:blur(0))]:backdrop-blur">
+    <span aria-hidden="true" className="font-mono text-sm text-blue-600">01</span>
+    <span className="ml-4 text-base font-medium text-slate-900"><span className="hidden lg:inline">Table of contents</span><span className="lg:hidden">Contents</span></span>
+    <button className="-mr-1 ml-auto flex h-8 w-8 items-center justify-center" aria-label="Toggle navigation menu" type="button" aria-expanded="false" data-headlessui-state="" id="headlessui-popover-button-:R9aemla:"><span className="absolute inset-0"></span>
+    <svg aria-hidden="true" fill="none" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" viewBox="0 0 24 24" className="h-6 w-6 stroke-slate-700">
+      <path d="m15 16-3 3-3-3M15 8l-3-3-3 3"></path></svg></button></div>
+      <div className="absolute inset-x-0 bottom-full z-10 h-4 bg-white"></div></div>
+      <div >
+        </div><div className="hidden sm:flex sm:h-32 sm:justify-center sm:border-b sm:border-slate-200 sm:bg-white/95 sm:[@supports(backdrop-filter:blur(0))]:bg-white/80 sm:[@supports(backdrop-filter:blur(0))]:backdrop-blur">
+        <ol role="list" className="mb-[-2px] grid auto-cols-[minmax(0,15rem)] grid-flow-col text-base font-medium text-slate-900 [counter-reset:section]">
+        <li className="flex [counter-increment:section]">
+            <a href="#paquetes" className="flex w-full flex-col items-center justify-center border-b-2 before:mb-2 before:font-mono before:text-sm before:content-[counter(section,decimal-leading-zero)] border-transparent before:text-slate-500 hover:bg-blue-50/40 hover:before:text-slate-900">
+                <span className="hidden lg:inline">Paquetes</span></a>
+                </li><li className="flex [counter-increment:section]">
+                    <a href="#contacto" className="flex w-full flex-col items-center justify-center border-b-2 before:mb-2 before:font-mono before:text-sm before:content-[counter(section,decimal-leading-zero)] border-transparent before:text-slate-500 hover:bg-blue-50/40 hover:before:text-slate-900">
+                      Contactanos</a></li>
+                    
+                    </ol>
+            
+            
+            
+            </div></div>
+          
 
-      <div className="mt-4 grid grid-cols-1 items-start gap-x-8 gap-y-8 sm:mt-16 sm:gap-y-16 lg:mt-24 lg:grid-cols-4">
-      <div className="relative -mx-4 flex overflow-x-auto pb-4 sm:mx-0 sm:block sm:overflow-visible sm:pb-0">
-      <div className="absolute bottom-0 left-0.5 top-2 hidden w-px bg-slate-200 lg:block"></div>
-      <div className="grid auto-cols-auto grid-flow-col justify-start gap-x-8 gap-y-10 whitespace-nowrap px-4 sm:mx-auto sm:max-w-2xl sm:grid-cols-3 sm:px-0 sm:text-center lg:grid-flow-row lg:grid-cols-1 lg:text-left" role="tablist" aria-orientation="vertical"><div className="relative lg:pl-8">
-        <svg aria-hidden="true" viewBox="0 0 6 6" className={`absolute left-[-0.5px] top-[0.5625rem] hidden h-1.5 w-1.5 overflow-visible lg:block fill-transparent ${!isActiveTodo ? 'fill-slate-500 stroke-slate-500' : 'fill-blue-600 stroke-blue-600' } `}>
-      <path d="M3 0L6 3L3 6L0 3Z" stroke-width="2" stroke-linejoin="round"></path>
-      </svg><div className="relative">
-        <div className={`font-mono text-sm ${!isActiveTodo ? 'text-slate-500' : 'text-blue-600' }`}>
-      <button onClick={() => filtro()} className="ui-not-focus-visible:outline-none" id="headlessui-tabs-tab-:R6cqlaqlla:" role="tab" type="button" aria-selected="false" tabindex="-1" data-headlessui-state="" aria-controls="headlessui-tabs-panel-:R3alaqlla:"><span className="absolute inset-0"></span>Todo tipo de dulce</button>
-      </div>
-      <time datetime="2022-04-04" className="mt-1.5 block text-2xl font-semibold tracking-tight text-blue-900">Todo</time>
-      </div>
-      </div>
-      <div className="relative lg:pl-8">
-        <svg aria-hidden="true" viewBox="0 0 6 6" className={`absolute left-[-0.5px] top-[0.5625rem] hidden h-1.5 w-1.5 overflow-visible lg:block fill-transparent ${!isActiveBotanas ? 'fill-slate-500 stroke-slate-500' : 'fill-blue-600 stroke-blue-600' } `}><path d="M3 0L6 3L3 6L0 3Z" stroke-width="2" stroke-linejoin="round">
-      </path>
-      </svg>
-      <div className="relative">
-        <div className={`font-mono text-sm ${!isActiveBotanas ? 'text-slate-500' : 'text-blue-600'} `}>
-          <button onClick={() => filtro('botanas')} className="ui-not-focus-visible:outline-none" id="headlessui-tabs-tab-:R6kqlaqlla:" role="tab" type="button" aria-selected="false" tabindex="-1" data-headlessui-state="" aria-controls="headlessui-tabs-panel-:R5alaqlla:">
-      <span className="absolute inset-0"></span>Salado</button>
-      </div>
-      <time datetime="2022-04-05" className="mt-1.5 block text-2xl font-semibold tracking-tight text-blue-900">Botanas</time>
-      </div>
-      </div>
-      <div className="relative lg:pl-8">
-      <svg aria-hidden="true" viewBox="0 0 6 6" className={`absolute left-[-0.5px] top-[0.5625rem] hidden h-1.5 w-1.5 overflow-visible lg:block fill-transparent ${!isActiveGomitas ? 'fill-slate-500 stroke-slate-500' : 'fill-blue-600 stroke-blue-600' } `}><path d="M3 0L6 3L3 6L0 3Z" stroke-width="2" stroke-linejoin="round">
-        </path>
-        </svg><div className="relative">
-        <div className={`font-mono text-sm ${!isActiveGomitas ? 'text-slate-500' : 'text-blue-600'} `}>
-          <button onClick={() => filtro('gomitas')} className="ui-not-focus-visible:outline-none" id="headlessui-tabs-tab-:R6sqlaqlla:" role="tab" type="button" aria-selected="false" tabindex="0" data-headlessui-state="selected" aria-controls="headlessui-tabs-panel-:R7alaqlla:">
-      <span className="absolute inset-0"></span>Dulce</button></div>
-      <time datetime="2022-04-06" className="mt-1.5 block text-2xl font-semibold tracking-tight text-blue-900">Gomitas</time>
-      </div>
-      </div>
-      <div className="relative lg:pl-8">
-      <svg aria-hidden="true" viewBox="0 0 6 6" className={`absolute left-[-0.5px] top-[0.5625rem] hidden h-1.5 w-1.5 overflow-visible lg:block fill-transparent ${!isActiveChocolate ? 'fill-slate-500 stroke-slate-500' : 'fill-blue-600 stroke-blue-600' } `}><path d="M3 0L6 3L3 6L0 3Z" stroke-width="2" stroke-linejoin="round"></path></svg><div className="relative">
-        <div className={`font-mono text-sm ${!isActiveChocolate ? 'text-slate-500' : 'text-blue-600'} `}>
-          <button onClick={() => filtro('chocolates')} className="ui-not-focus-visible:outline-none" id="headlessui-tabs-tab-:R6sqlaqlla:" role="tab" type="button" aria-selected={isActiveChocolate} tabindex="0" data-headlessui-state="" aria-controls="headlessui-tabs-panel-:R7alaqlla:">
-      <span className="absolute inset-0"></span>Dulce y Salado</button></div>
-      <time datetime="2022-04-06" className="mt-1.5 block text-2xl font-semibold tracking-tight text-blue-900">Chocolates</time>
-      </div>
-      </div>
-      </div>
-      </div>
-      <div className="lg:col-span-3">
+
+ 
+
+<section id="paquetes" aria-labelledby="pricing-title" className=" bg-body-tertiary py-20 sm:py-32">
+  <div className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8"><div className="mx-auto max-w-2xl text-center"><h2 id="pricing-title" className="text-3xl font-medium tracking-tight text-gray-900">Flat pricing, no management fees.</h2><p className="mt-2 text-lg text-gray-600">Whether you’re one person trying to get ahead or a big firm trying to take over the world, we’ve got a plan for you.</p></div><div className="mt-8 flex justify-center"><div className="relative"><div className="grid grid-cols-2" id="headlessui-radiogroup-:Rraqlla:" role="radiogroup"><div className="cursor-pointer border border-gray-300 px-[calc(theme(spacing.3)-1px)] py-[calc(theme(spacing.2)-1px)] text-sm text-gray-700 outline-2 outline-offset-2 transition-colors hover:border-gray-400 rounded-l-lg" id="headlessui-radiogroup-option-:Rqraqlla:" role="radio" aria-checked="true" tabindex="0" data-headlessui-state="checked">Monthly</div><div className="cursor-pointer border border-gray-300 px-[calc(theme(spacing.3)-1px)] py-[calc(theme(spacing.2)-1px)] text-sm text-gray-700 outline-2 outline-offset-2 transition-colors hover:border-gray-400 -ml-px rounded-r-lg" id="headlessui-radiogroup-option-:R1araqlla:" role="radio" aria-checked="false" tabindex="-1" data-headlessui-state="">Annually</div></div><div aria-hidden="true" className="pointer-events-none absolute inset-0 z-10 grid grid-cols-2 overflow-hidden rounded-lg bg-cyan-500 transition-all duration-300 [clip-path:inset(0_50%_0_0)]"><div className="py-2 text-center text-sm font-semibold text-white">Monthly</div><div className="py-2 text-center text-sm font-semibold text-white -ml-px">Annually</div></div></div></div><div className="mx-auto mt-16 grid max-w-2xl grid-cols-1 items-start gap-x-8 gap-y-10 sm:mt-20 lg:max-w-none lg:grid-cols-3"><section className="flex flex-col overflow-hidden rounded-3xl p-6 shadow-lg shadow-gray-900/5 bg-white"><h3 className="flex items-center text-sm font-semibold text-gray-900"><svg viewBox="0 0 40 40" aria-hidden="true" className="h-6 w-6 flex-none fill-gray-300"><path fill-rule="evenodd" clip-rule="evenodd" d="M20 40C8.954 40 0 31.046 0 20S8.954 0 20 0s20 8.954 20 20-8.954 20-20 20ZM4 20c0 7.264 5.163 13.321 12.02 14.704C17.642 35.03 19 33.657 19 32V8c0-1.657-1.357-3.031-2.98-2.704C9.162 6.68 4 12.736 4 20Z"></path></svg><span className="ml-4">Starter</span></h3><p className="relative mt-5 flex text-3xl tracking-tight text-gray-900">$0</p><p className="mt-3 text-sm text-gray-700">You’re new to investing but want to do it right. Get started for free.</p><div className="order-last mt-6"><ul role="list" className="-my-2 divide-y text-sm divide-gray-200 text-gray-700"><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-cyan-500"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">Commission-free trading</span></li><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-cyan-500"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">Multi-layered encryption</span></li><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-cyan-500"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">One tip every day</span></li><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-cyan-500"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">Invest up to $1,500 each month</span></li></ul></div><a className="inline-flex justify-center rounded-lg py-2 px-3 text-sm font-semibold outline-2 outline-offset-2 transition-colors bg-gray-800 text-white hover:bg-gray-900 active:bg-gray-800 active:text-white/80 mt-6" aria-label="Get started with the Starter plan for [object Object]" href="/register">Get started for free</a></section><section className="flex flex-col overflow-hidden rounded-3xl p-6 shadow-lg shadow-gray-900/5 bg-white"><h3 className="flex items-center text-sm font-semibold text-gray-900"><svg viewBox="0 0 40 40" aria-hidden="true" className="h-6 w-6 flex-none fill-gray-500"><path fill-rule="evenodd" clip-rule="evenodd" d="M20 40C8.954 40 0 31.046 0 20S8.954 0 20 0s20 8.954 20 20-8.954 20-20 20ZM4 20c0 7.264 5.163 13.321 12.02 14.704C17.642 35.03 19 33.657 19 32V8c0-1.657-1.357-3.031-2.98-2.704C9.162 6.68 4 12.736 4 20Z"></path></svg><span className="ml-4">Investor</span></h3><p className="relative mt-5 flex text-3xl tracking-tight text-gray-900"><span aria-hidden="false" className="transition duration-300">$7</span><span aria-hidden="true" className="absolute left-0 top-0 transition duration-300 pointer-events-none -translate-x-6 select-none opacity-0">$70</span></p><p className="mt-3 text-sm text-gray-700">You’ve been investing for a while. Invest more and grow your wealth faster.</p><div className="order-last mt-6"><ul role="list" className="-my-2 divide-y text-sm divide-gray-200 text-gray-700"><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-cyan-500"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">Commission-free trading</span></li><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-cyan-500"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">Multi-layered encryption</span></li><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-cyan-500"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">One tip every hour</span></li><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-cyan-500"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">Invest up to $15,000 each month</span></li><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-cyan-500"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">Basic transaction anonymization</span></li></ul></div><a className="inline-flex justify-center rounded-lg py-2 px-3 text-sm font-semibold outline-2 outline-offset-2 transition-colors bg-gray-800 text-white hover:bg-gray-900 active:bg-gray-800 active:text-white/80 mt-6" aria-label="Get started with the Investor plan for [object Object]" href="/register">Subscribe</a></section><section className="flex flex-col overflow-hidden rounded-3xl p-6 shadow-lg shadow-gray-900/5 order-first bg-gray-900 lg:order-none"><h3 className="flex items-center text-sm font-semibold text-white"><svg viewBox="0 0 40 40" aria-hidden="true" className="h-6 w-6 flex-none fill-cyan-500"><path fill-rule="evenodd" clip-rule="evenodd" d="M20 40C8.954 40 0 31.046 0 20S8.954 0 20 0s20 8.954 20 20-8.954 20-20 20ZM4 20c0 7.264 5.163 13.321 12.02 14.704C17.642 35.03 19 33.657 19 32V8c0-1.657-1.357-3.031-2.98-2.704C9.162 6.68 4 12.736 4 20Z"></path></svg><span className="ml-4">VIP</span></h3><p className="relative mt-5 flex text-3xl tracking-tight text-white"><span aria-hidden="false" className="transition duration-300">$199</span><span aria-hidden="true" className="absolute left-0 top-0 transition duration-300 pointer-events-none -translate-x-6 select-none opacity-0">$1,990</span></p><p className="mt-3 text-sm text-gray-300">You’ve got a huge amount of assets but it’s not enough. To the moon.</p><div className="order-last mt-6"><ul role="list" className="-my-2 divide-y text-sm divide-gray-800 text-gray-300"><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-white"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">Commission-free trading</span></li><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-white"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">Multi-layered encryption</span></li><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-white"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">Real-time tip notifications</span></li><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-white"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">No investment limits</span></li><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-white"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">Advanced transaction anonymization</span></li><li className="flex py-2"><svg viewBox="0 0 24 24" aria-hidden="true" className="h-6 w-6 flex-none text-white"><path d="M9.307 12.248a.75.75 0 1 0-1.114 1.004l1.114-1.004ZM11 15.25l-.557.502a.75.75 0 0 0 1.15-.043L11 15.25Zm4.844-5.041a.75.75 0 0 0-1.188-.918l1.188.918Zm-7.651 3.043 2.25 2.5 1.114-1.004-2.25-2.5-1.114 1.004Zm3.4 2.457 4.25-5.5-1.187-.918-4.25 5.5 1.188.918Z" fill="currentColor"></path><circle cx="12" cy="12" r="8.25" fill="none" stroke="currentColor" stroke-width="1.5" stroke-linecap="round" stroke-linejoin="round"></circle></svg><span className="ml-4">Automated tax-loss harvesting</span></li></ul></div><a className="inline-flex justify-center rounded-lg py-2 px-3 text-sm font-semibold outline-2 outline-offset-2 transition-colors relative overflow-hidden bg-cyan-500 text-white before:absolute before:inset-0 active:before:bg-transparent hover:before:bg-white/10 active:bg-cyan-600 active:text-white/80 before:transition-colors mt-6" aria-label="Get started with the VIP plan for [object Object]" href="/register">Subscribe</a></section></div></div></section>
+
+
+
+
+
+<section id="contacto" aria-labelledby="author-title" className="relative scroll-mt-14 pb-3 pt-8 sm:scroll-mt-32 sm:pb-16 sm:pt-10 lg:pt-16">
+  <div className="absolute inset-x-0 bottom-0 top-1/2 text-slate-900/10 [mask-image:linear-gradient(transparent,white)]">
+    <svg aria-hidden="true" className="absolute inset-0 h-full w-full"><defs>
+      <pattern id=":S7:" width="128" height="128" patternUnits="userSpaceOnUse" x="50%" y="100%">
+        <path d="M0 128V.5H128" fill="none" stroke="currentColor"></path></pattern></defs><rect width="100%" height="100%" fill="url(#:S7:)"></rect></svg></div>
+        <div className="relative mx-auto max-w-5xl pt-16 sm:px-6">
+          <div className="bg-slate-50 pt-px sm:rounded-6xl">
+            <div className="relative mx-auto -mt-16 h-44 w-44 overflow-hidden rounded-full bg-slate-200 md:float-right md:h-64 md:w-64 md:[shape-outside:circle(40%)] lg:mr-20 lg:h-72 lg:w-72">
+              <img alt="" /></div>
+              <div className="px-4 py-10 sm:px-10 sm:py-16 md:py-20 lg:px-20 lg:py-32">
+                <h2 className="inline-flex items-center rounded-full px-4 py-1 text-blue-600 ring-1 ring-inset ring-blue-600" id="author-title">
+                  <span className="font-mono text-sm" aria-hidden="true">05</span><span className="ml-3 h-3.5 w-px bg-blue-600/20"></span>
+                  <span className="ml-3 text-base font-medium tracking-tight">Author</span></h2>
+                  <p className="mt-8 font-display text-5xl font-extrabold tracking-tight text-slate-900 sm:text-6xl">
+                    <span className="block text-blue-600">Mira Lindehoff –</span> Hey there, I’m the author behind ‘Everything Starts as a Square’.</p>
+                    <p className="mt-4 text-lg tracking-tight text-slate-700">I’ve been designing icons professionally for over a decade and have worked with dozens of the biggest brands to create custom sets for their products. I’m an accomplished conference speaker, and have been teaching icon design workshops every month for the last three years. I’ve worked with designers of all skill levels and honed my way of teaching to really click for anyone who has the itch to start designing their own icons.</p><p className="mt-8"><a className="inline-flex items-center text-base font-medium tracking-tight text-blue-600" href="#"><svg aria-hidden="true" viewBox="0 0 40 40" className="h-10 w-10 fill-current"><path d="M13.817 33.753c12.579 0 19.459-10.422 19.459-19.458 0-.297 0-.592-.02-.884a13.913 13.913 0 0 0 3.411-3.543 13.65 13.65 0 0 1-3.928 1.077 6.864 6.864 0 0 0 3.007-3.784 13.707 13.707 0 0 1-4.342 1.66 6.845 6.845 0 0 0-11.655 6.239A19.417 19.417 0 0 1 5.654 7.915a6.843 6.843 0 0 0 2.117 9.128 6.786 6.786 0 0 1-3.104-.853v.086a6.842 6.842 0 0 0 5.487 6.704 6.825 6.825 0 0 1-3.088.116 6.847 6.847 0 0 0 6.39 4.75A13.721 13.721 0 0 1 3.334 30.68a19.36 19.36 0 0 0 10.483 3.066"></path></svg><span className="ml-4">Follow on Twitter</span></a></p></div></div></div></section>
+
+
+
+
+
+
+
+                    </div>
     
-   <div className="grid grid-cols-1 gap-x-8 gap-y-10 ui-not-focus-visible:outline-none sm:grid-cols-2 sm:gap-y-16 md:grid-cols-3" id="headlessui-tabs-panel-:R7alaqlla:" role="tabpanel" tabindex="0"  data-headlessui-state="selected" aria-labelledby="headlessui-tabs-tab-:R6sqlaqlla:">
-    {
-   renderView()
-    }
-   
-        </div>
-        
-        </div></div></div>
-
-
-
-
-
-      {/* <div className=' flex mt-5 w-full items-center justify-center'>
-        <input type='text' placeholder='Buscar producto' className='w-80 rounded-lg border border-green-200 shadow p-4 mb-4 focus:outline-none' 
-          onChange={search}></input>
-      </div> */}
-      <div className='container grid mt-3 '>
-        { //renderizado  gap-4 mt-4 grid-cols-2 sm:grid-cols-4 w-full max-w-screen-lg px-2
-
-   
-         
-
-     
-        }
-
-
-      </div>
-      {openModal && (
-        <ProductModal>
-          <ProductDetail></ProductDetail>
-        </ProductModal>
-      )}
-      <CheckoutSideMenu></CheckoutSideMenu>
-    </div>
-  )
+    )
 }
-
-export default Home
+export default HomePage
